requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorthief==0.2.1
cryptography==41.0.7
databases==0.8.0
ecdsa==0.18.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
pillow==10.2.0
psycopg2-binary==2.9.9
pyairtable==2.2.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode-artistic==3.0.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
segno==1.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
